Refactoring: Separate import::csv module into sub modules.
Existing code has low coverage.
Separation will enforce adding more tests.
